Kill ClassNamesSupport
